fix: Policy CleanUpAfterDeletion prevents kubernetes CustomResource deletion when pulsar resources are already deleted
Deleting a resource such as PulsarTopic with policy CleanUpAfterDeletion sometimes results in the resource getting stuck. The operator will log: "error":"code: 404 reason: Topic not exist"
Cause: During the reconciliation loop the deletion of the pulsar resource succeeds but removing the finalizer can fail due to updates to the kubernetes resource resulting in following error: "Operation cannot be fulfilled on pulsartopics.resource.streamnative.io \"xxx\": the object has been modified; please apply your changes to the latest version and try again"
Other than resource updates this can also happen when there is a network hickup causing the call to remove the finalizer to fail. So this cannot be prevented.
Fix: If the pulsar resource no longer exists, it is safe to remove the finalizer. It seems like this fix was already attempted for some resources but was inverted.